Refactor k8s workload attestor to support windows attestation
