+++ pkg/agent/plugin/nodeattestor/k8ssat/sat_posix.go
@@ -1,0 +1,8 @@
+//go:build !windows
+// +build !windows
+
+package sat
+
+func getDefaultTokenPath() string {
+	return defaultTokenPath
+}
