+++ pkg/agent/plugin/nodeattestor/k8ssat/sat_windows.go
@@ -1,0 +1,21 @@
+//go:build windows
+// +build windows
+
+package sat
+
+import (
+	"os"
+	"path"
+)
+
+const (
+	containerMountPoint = "CONTAINER_SANDBOX_MOUNT_POINT"
+)
+
+func getDefaultTokenPath() string {
+	mountPoint := os.Getenv(containerMountPoint)
+	if mountPoint == "" {
+		return defaultTokenPath
+	}
+	return path.Join(mountPoint, defaultTokenPath)
+}
