+++ pkg/agent/plugin/workloadattestor/k8s/k8s_windows.go
@@ -4,31 +4,46 @@
 package k8s
 
 import (
-	"context"
-
-	workloadattestorv1 "github.com/spiffe/spire-plugin-sdk/proto/spire/plugin/agent/workloadattestor/v1"
-	configv1 "github.com/spiffe/spire-plugin-sdk/proto/spire/service/common/config/v1"
-	"github.com/spiffe/spire/pkg/common/catalog"
+	"github.com/hashicorp/go-hclog"
+	"github.com/spiffe/spire/pkg/common/container/process"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
+	"k8s.io/apimachinery/pkg/types"
 )
 
-type Plugin struct {
-	workloadattestorv1.UnimplementedWorkloadAttestorServer
-	configv1.UnsafeConfigServer
+const (
+	containerMountPoint = "CONTAINER_SANDBOX_MOUNT_POINT"
+)
+
+func createHelper(c *Plugin) (ContainerHelper, error) {
+	return &containerHelper{
+		ph: process.CreateHelper(),
+	}, nil
 }
 
-func builtin(p *Plugin) catalog.BuiltIn {
-	return catalog.MakeBuiltIn(pluginName,
-		workloadattestorv1.WorkloadAttestorPluginServer(p),
-		configv1.ConfigServiceServer(p),
-	)
+type containerHelper struct {
+	ph process.Helper
 }
 
-func New() *Plugin {
-	return &Plugin{}
+func (h *containerHelper) GetPodUIDAndContainerID(pID int32, log hclog.Logger) (types.UID, string, error) {
+	containerID, err := h.ph.GetContainerIDByProcess(pID, log)
+	if err != nil {
+		return types.UID(""), "", status.Errorf(codes.Internal, "failed to get container ID: %v", err)
+	}
+
+	return types.UID(""), containerID, nil
 }
 
-func (p *Plugin) Configure(ctx context.Context, req *configv1.ConfigureRequest) (*configv1.ConfigureResponse, error) {
-	return nil, status.Error(codes.Unimplemented, "plugin not supported in this platform")
+func (p *Plugin) defaultKubeletCAPath() string {
+	mountPoint := p.getenv(containerMountPoint)
+	return mountPoint + defaultKubeletCAPath
+}
+
+func (p *Plugin) defaultTokenPath() string {
+	mountPoint := p.getenv(containerMountPoint)
+	return mountPoint + defaultTokenPath
+}
+
+func isNotPod(itemPodUID, podUID types.UID) bool {
+	return false
 }
